+++ components/com_resources/controllers/attachments.php
@@ -931,6 +931,10 @@ class ResourcesControllerAttachments extends Hubzero_Controller
 		// Build the path
 		$path = $this->_buildUploadPath($listdir, '');
 
+		$base   = JPATH_ROOT . '/' . trim($this->config->get('webpath', '/site/resources'), '/');
+		$baseY = $base . '/'. JFactory::getDate($row->created)->format("Y");
+		$baseM = $baseY . '/' . JFactory::getDate($row->created)->format("m");
+
 		// Check if the folder even exists
 		if (!file_exists($path) or !$path) 
 		{
@@ -938,10 +942,19 @@ class ResourcesControllerAttachments extends Hubzero_Controller
 		} 
 		else 
 		{
-			// Attempt to delete the folder
-			if (!JFile::delete($path)) 
+			if ($path == $base 
+			 || $path == $baseY
+			 || $path == $baseM)
 			{
-				$this->setError(JText::_('COM_CONTRIBUTE_UNABLE_TO_DELETE_FILE'));
+				$this->setError(JText::_('Invalid directory.'));
+			}
+			else
+			{
+				// Attempt to delete the folder
+				if (!JFile::delete($path)) 
+				{
+					$this->setError(JText::_('COM_CONTRIBUTE_UNABLE_TO_DELETE_FILE'));
+				}
 			}
 		}
