+++ components/com_resources/controllers/create.php
@@ -1563,6 +1563,10 @@ class ResourcesControllerCreate extends Hubzero_Controller
 				// Build the path
 				$path = $this->_buildUploadPath($listdir, '');
 
+				$base   = JPATH_ROOT . '/' . trim($this->config->get('webpath', '/site/resources'), '/');
+				$baseY = $base . '/'. JFactory::getDate($child->created)->format("Y");
+				$baseM = $baseY . '/' . JFactory::getDate($child->created)->format("m");
+
 				// Check if the folder even exists
 				if (!is_dir($path) or !$path) 
 				{
@@ -1570,10 +1574,19 @@ class ResourcesControllerCreate extends Hubzero_Controller
 				} 
 				else 
 				{
-					// Attempt to delete the folder
-					if (!JFolder::delete($path)) 
+					if ($path == $base 
+					 || $path == $baseY
+					 || $path == $baseM)
 					{
-						$this->setError(JText::_('COM_CONTRIBUTE_UNABLE_TO_DELETE_DIRECTORY'));
+						$this->setError(JText::_('Invalid directory.'));
+					}
+					else
+					{
+						// Attempt to delete the folder
+						if (!JFolder::delete($path)) 
+						{
+							$this->setError(JText::_('COM_CONTRIBUTE_UNABLE_TO_DELETE_DIRECTORY'));
+						}
 					}
 				}
